refactor(hub-common): move all hubSearchItems helpers into their own file for better testability
affects: @esri/hub-common
