+++ packages/common/src/search/_internal/hubSearchItemsHelpers/getQueryString.ts
@@ -1,0 +1,8 @@
+export function getQueryString(queryParams: Record<string, any>) {
+  const result = Object.entries(queryParams)
+    .filter(([_key, value]) => !!value)
+    .map(([key, value]) => `${key}=${value}`)
+    .join("&");
+
+  return result && `?${result}`;
+}
