+++ packages/common/test/search/_internal/portalSearchItems.test.ts
@@ -1,11 +1,5 @@
 import * as Portal from "@esri/arcgis-rest-portal";
-import {
-  cloneObject,
-  IFilter,
-  IHubSearchOptions,
-  IPredicate,
-  IQuery,
-} from "../../../src";
+import { cloneObject, IHubSearchOptions, IQuery } from "../../../src";
 
 import * as SimpleResponse from "../../mocks/portal-search/simple-response.json";
 import * as AllTypesResponse from "../../mocks/portal-search/response-with-key-types.json";
@@ -15,12 +9,6 @@ import {
   portalSearchItems,
   WellKnownItemPredicates,
 } from "../../../src/search/_internal/portalSearchItems";
-import {
-  formatFilterBlock,
-  formatPredicate,
-  getOgcItemQueryParams,
-} from "../../../src/search/_internal/hubSearchItems";
-import { UserSession } from "@esri/arcgis-rest-auth";
 
 describe("portalSearchItems Module:", () => {
   describe("portalSearchItems:", () => {
